Future concept for flight-ready weight reduction
NOTE: This is a rough concept ONLY. Design idea for material reduction and overall simplification for next generation of the PSAS electric feed system. Driving the impeller directly from an extended motor shaft removes the need for a coupler and reduces the footprint of the system. Excess material can be removed from the pump casing as well, further reducing the weight.